+++ app/state/actions/DataActions.js
@@ -8,9 +8,6 @@ import sampleData from './sampleData';
 
 // A bit ugly, but we'll just hardcode for now.
 function createArcGISUrl(dataset, doCountOnly, offset, count) {
-  // const fields = "date_opened,record_status,record_status_date,"
-  //              + "record_type,record_type_group,record_type_category,record_type_type,"
-  //              + "record_type_subtype,latitude,longitude";
   const q = dataset.query;
   let url = q.url + `?where=${encodeURIComponent(q.where)}&text=&objectIds=&time=&geometry=`;
   url += "&geometryType=esriGeometryEnvelope&inSR=&spatialRel=esriSpatialRelIntersects";
@@ -26,27 +23,17 @@ function createArcGISUrl(dataset, doCountOnly, offset, count) {
   return url;
 }
 
-function translateModule (record_type_group) {
-  switch(record_type_group) {
-    case 'Permits':
-      return 'Construction';
-    case 'Planning':
-      return 'Planning';
-    case 'Services':
-      return 'Enforcement';
-    default:
-      return 'Unknown';
-  }
-}
-
-function transformRecord(inR0, useAttributes) {
-  let inR = useAttributes?inR0.attributes:inR0;
+// FIXTHIS if production:
+//  This should be driven by the query.fields configuration parameter.
+function transformRecord(inR0, usingBackupServer) {
+  // The SimpliCity API puts attributes in a sub-object.
+  let inR = usingBackupServer?inR0.attributes:inR0;
   let record = {
     name: inR.record_name,
     date_opened: inR.date_opened,
     record_status: inR.record_status,
     record_status_date: inR.record_status_date,
-    record_type_group: translateModule(inR.record_type_group),
+    record_type_group: inR.record_type_group,
     record_type_type: inR.record_type_type,
     record_type_category: inR.record_type_category,
     record_type: inR.record_type,
@@ -80,8 +67,9 @@ function doFakeFetch (dataset, dispatch, maxRecordCount, count) {
 }
 
 function doApiFetch (dataset, dispatch, maxRecordCount, count) {
-  // Initial call gets the expected record count so that we can keep the user informed
-  // as to what's happening.
+  // Initial call gets the expected record count so we could keep the user informed
+  // as to where we are in the process. Except that right now the API is returning
+  // more than expected, so not trying to tell the user about it yet.
   let url, initialCall = (count < 0);
   if (initialCall)
     url = createArcGISUrl(dataset, true, count, maxRecordCount);
@@ -116,11 +104,11 @@ function doApiFetch (dataset, dispatch, maxRecordCount, count) {
           }
         }
         else {
-          const maxRecords = 5000; // DEBUG
+          const maxRecords = 30000; // DEBUG
           console.log("Received records of length " + json.features.length);
           const done = json.features.length < maxRecordCount || count >= maxRecords;
           const data = json.features.map( (item) => {
-            return transformRecord(item, dataset.query.use_attributes);
+            return transformRecord(item, dataset.query.usingBackupServer);
           });
 
           dispatch({type: actionTypes.UPDATE_DATASET, data: {
@@ -148,19 +136,18 @@ function doApiFetch (dataset, dispatch, maxRecordCount, count) {
     });
 }
 
+///////////////////////////////////////////////////////////////////
+// Note for the future:
+//  I've got a lot of hardcoded stuff in this file. In a real
+//  version, we'd break out a set of different loaders and make
+//  them all configurable.
+///////////////////////////////////////////////////////////////////
+
 export function fetchDataset(dataset, dispatch) {
   dispatch ({ type: actionTypes.FETCH_DATASET, data: { dataset } }); // Just registers the dataset
 
-  /******* ProjectCommentary *********
-      I am constructing the query here for this test project, however,
-      in a production version we'd probably have a set of dataloader helpers
-      that were invoked based on the dataset source_type that also understood how
-      to process different dataset_types for use here.
-  */
-
   if (dataset.source_type == 'arcgis-rest') {
     const maxRecordCount = dataset.query.maxRecordCount;
-    // Maybe do config object rather than args here??????????
     if (dataset.simulate_api) {
       return doFakeFetch (dataset, dispatch, maxRecordCount, -1);
     }
